// src/redux/types/assignedTasksTypes.tsx
export enum ASSIGNED_TASK_TYPES {
  ASSIGN_TASK_FETCH = "ASSIGN_TASK_FETCH",
  ASSIGN_TASK_LOADING = "ASSIGN_TASK_LOADING",
  ASSIGN_TASK_FAILED = "ASSIGN_TASK_FAILED",
  ASSIGN_TASK_SUCCESS = "ASSIGN_TASK_SUCCESS",
  ASSIGNED_TASKS_FETCH = "ASSIGNED_TASKS_FETCH",
  ASSIGNED_TASKS_LOADING = "ASSIGNED_TASKS_LOADING",
  ASSIGNED_TASKS_FAILED = "ASSIGNED_TASKS_FAILED",
  ASSIGNED_TASKS_SUCCESS = "ASSIGNED_TASKS_SUCCESS",
  MARK_ASSIGNED_TASK_COMPLETED_FETCH = "MARK_ASSIGNED_TASK_COMPLETED_FETCH",
  MARK_ASSIGNED_TASK_CANCELLED_FETCH = "MARK_ASSIGNED_TASK_CANCELLED_FETCH",
  EDIT_ASSIGNED_TASK_FETCH = "EDIT_ASSIGNED_TASK_FETCH",
  EDIT_ASSIGNED_TASK_LOADING = "EDIT_ASSIGNED_TASK_LOADING",
  EDIT_ASSIGNED_TASK_SUCCESS = "EDIT_ASSIGNED_TASK_SUCCESS",
  EDIT_ASSIGNED_TASK_FAILED = "EDIT_ASSIGNED_TASK_FAILED",
  RECIEVED_TASKS_SUCCESS = "RECIEVED_TASKS_SUCCESS",
  RECIEVED_TASKS_FETCH = "RECIEVED_TASKS_FETCH",
  RECIEVED_TASKS_LOADING = "RECIEVED_TASKS_LOADING",
  RECIEVED_TASKS_FAILED = "RECIEVED_TASKS_FAILED",
}